feat: Remove deprecated timer and stopwatch components and restructure imports
