+++ apps/frontend/app/components/fitness.action/stat-display-and-input.tsx
